+++ CHANGELOG.md
@@ -1,4 +1,15 @@
-CHANGELOG
+# CHANGELOG
+
+v3.2.0 (2017-01-20)
+-------------------
+
+### New Feature
+* Add `setup` recipe to install both: WordPress (see `install`) and the WordPress core (see `install_core`) via WordPress CLI.
+* Add Vagrantfiles (https://www.vagrantup.com/) for Virtualbox (`vagrant-virtualbox`) and Amazon EC2 (`vagrant-aws`)
+
+### Bug
+* Fix reprovisioning after VM restart (observed on Ubuntu 14.04)
+* Fix to enable wordpress apache config
 
 v3.0.0 (2015-02-24)
 -------------------